app/requirements.txt
requests
python-telegram-bot
firebase
firebase-admin